+++ resources/views/dashboard/reps/all.blade.php
@@ -11,15 +11,20 @@
   
     <div class="btn-group">
         
-        <a href="{{route('admin.rep.add')}}" class="btn btn-warning"> 
-            <i class="fa fa-plus"></i>  @lang('site.add') </a> 
+        @if(has_permission('reps_add'))
+            <a href="{{route('admin.rep.add')}}" class="btn btn-warning"> 
+                <i class="fa fa-plus"></i>  @lang('site.add') </a> 
+        @endif
 
-        <a href="{{route('export.excel.users')}}" class="btn btn-success"> 
-            <i class="fa fa-download"></i>  @lang('site.excel') </a> 
+        @if(has_permission('reps_show'))
+            <a href="{{route('export.excel.users')}}" class="btn btn-success"> 
+                <i class="fa fa-download"></i>  @lang('site.excel') </a> 
+        @endif
 
-        <a href="{{route('export.pdf.users')}}" class="btn btn-info"> 
-            <i class="fa fa-file"></i>  @lang('site.pdf') </a> 
-         
+        @if(has_permission('reps_show'))
+            <a href="{{route('export.pdf.users')}}" class="btn btn-info"> 
+                <i class="fa fa-file"></i>  @lang('site.pdf') </a> 
+        @endif
     </div>
 
 <br/> <br/>
@@ -28,13 +33,23 @@
     <thead>
     <tr class="headings">
         <th>#  </th>
+        
         <th> @lang('site.user_id')</th>
+        
         <th> <i class="fa fa-camera"> </i> </th>
+        
         <th> @lang('site.name')   </th>                          
+        
         <th> @lang('site.vip') </th>
+        
+        <th> @lang('site.user_role') </th>
+        
         <th> @lang('site.active') </th>
+        
         <th> @lang('site.saudi') </th>
+        
         <th> @lang('site.rating') </th>
+        
         <th style="width:120px;"></th>
     </tr>
     </thead>
@@ -54,7 +69,7 @@
             </td>
 
             <td>{{$item->name}}</td>
-             
+
             <td>
                 @if($item->vip ==1) <button class="btn btn-success btn-xs">
                          <i class="fa fa-check"></i> @lang('site.yes') </button>
@@ -65,12 +80,18 @@
             </td>
 
             <td>
-                @if($item->active ==1) <button class="btn btn-success btn-xs">
-                    <i class="fa fa-check"></i> @lang('site.yes') </button>
+                @foreach ($item->rep_roles as $rep_role)
+                    <label class="btn btn-default"> {{ __($rep_role->role['name_'.my_lang()]) }} </label>
+                @endforeach
+            </td>
+            
+            <td>
+                @if($item->active ==1) <button class="btn btn-success btn-xs" onclick="activate({{ $item->id }})">
+                    <i class="fa fa-check"></i> @lang('site.de_activate') </button>
                 @else
-                    <button class="btn btn-warning btn-xs">
-                    <i class="fa fa-close"></i> @lang('site.no') </button>
-                @endif     
+                    <button class="btn btn-warning btn-xs" onclick="activate({{ $item->id }})">
+                    <i class="fa fa-close"></i> @lang('site.activate') </button>
+                @endif      
             </td>
 
 
@@ -90,9 +111,13 @@
                     {{ setting('whatsapp_msg') }}"> <i class="fa fa-whatsapp"></i>
                 </a>
 
-                <a href="{{ url('admin/rep',$item->id) }}" class="btn btn-info btn-xs"> <i class="fa fa-edit"></i> </a>
+                @if(has_permission('reps_edit'))
+                    <a href="{{ url('admin/rep',$item->id) }}" class="btn btn-info btn-xs"> <i class="fa fa-edit"></i> </a>
+                @endif
 
-                <a onclick="deleteItem({{ $item->id }})" class="btn btn-danger btn-xs"><i class="fa fa-trash"></i> </a>
+                @if(has_permission('reps_delete'))
+                    <a onclick="deleteItem({{ $item->id }})" class="btn btn-danger btn-xs"><i class="fa fa-trash"></i> </a>
+                @endif
             </td>
         </tr>
            
@@ -112,5 +137,7 @@
     @include('dashboard.layouts.message_growl') 
 
     @include('dashboard.ajax.delete',['target'=>'rep']) 
+
+    @include('dashboard.ajax.activate',['target'=>'rep']) 
  
 @endsection
