+++ resources/views/dashboard/users/all.blade.php
@@ -11,15 +11,20 @@
   
     <div class="btn-group">
         
-        <a href="{{route('admin.user.add')}}" class="btn btn-warning"> 
-            <i class="fa fa-plus"></i>  @lang('site.add') </a> 
+        @if(has_permission('users_add'))
+            <a href="{{route('admin.user.add')}}" class="btn btn-warning"> 
+                <i class="fa fa-plus"></i>  @lang('site.add') </a> 
+        @endif
 
-        <a href="{{route('export.excel.users')}}" class="btn btn-success"> 
-            <i class="fa fa-download"></i>  @lang('site.excel') </a> 
+        @if(has_permission('users_show'))
+            <a href="{{route('export.excel.users')}}" class="btn btn-success"> 
+                <i class="fa fa-download"></i>  @lang('site.excel') </a> 
+        @endif
 
-        <a href="{{route('export.pdf.users')}}" class="btn btn-info"> 
-            <i class="fa fa-file"></i>  @lang('site.pdf') </a> 
-         
+        @if(has_permission('users_show'))
+            <a href="{{route('export.pdf.users')}}" class="btn btn-info"> 
+                <i class="fa fa-file"></i>  @lang('site.pdf') </a> 
+        @endif 
     </div>
 
 <br/> <br/>
@@ -62,13 +67,13 @@
                 @endif 
             </td>
 
-            <td>
-                @if($item->active ==1) <button class="btn btn-success btn-xs">
-                    <i class="fa fa-check"></i> @lang('site.yes') </button>
+            <td>                
+                @if($item->active ==1) <button class="btn btn-success btn-xs" onclick="activate({{ $item->id }})">
+                    <i class="fa fa-check"></i> @lang('site.de_activate') </button>
                 @else
-                    <button class="btn btn-warning btn-xs">
-                    <i class="fa fa-close"></i> @lang('site.no') </button>
-                @endif     
+                    <button class="btn btn-warning btn-xs" onclick="activate({{ $item->id }})">
+                    <i class="fa fa-close"></i> @lang('site.activate') </button>
+                @endif                     
             </td>
 
 
@@ -82,15 +87,19 @@
             </td>
           
             <td> {{ $item->available_requests }} </td>
-    
+     
             <td>
                 <a class="whatsapp btn btn-success btn-xs" target="_blank" href="https://wa.me/966{{$item->mobile}}?text=
                     {{ setting('whatsapp_msg') }}"> <i class="fa fa-whatsapp"></i>
                 </a>
 
-                <a href="{{ url('admin/user',$item->id) }}" class="btn btn-info btn-xs"> <i class="fa fa-edit"></i> </a>
+                @if(has_permission('users_edit'))
+                    <a href="{{ url('admin/user',$item->id) }}" class="btn btn-info btn-xs"> <i class="fa fa-edit"></i> </a>
+                @endif
 
-                <a onclick="deleteItem({{ $item->id }})" class="btn btn-danger btn-xs"><i class="fa fa-trash"></i> </a>
+                @if(has_permission('users_delete'))
+                    <a onclick="deleteItem({{ $item->id }})" class="btn btn-danger btn-xs"><i class="fa fa-trash"></i> </a>
+                @endif
             </td>
         </tr>
            
@@ -110,5 +119,6 @@
     @include('dashboard.layouts.message_growl') 
 
     @include('dashboard.ajax.delete',['target'=>'user']) 
+    @include('dashboard.ajax.activate',['target'=>'user']) 
  
 @endsection
